Move recipe to component
